# docs/ai-system/OPENAPI_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: CoinDaily AI System API
  version: 1.0.0
  description: |
    Comprehensive REST API documentation for CoinDaily's AI System.
    
    ## Overview
    The CoinDaily AI System provides a comprehensive suite of AI-powered features for cryptocurrency content management,
    market analysis, social media automation, and more.
    
    ## Base URL
    - Production: `https://api.coindaily.com`
    - Staging: `https://staging-api.coindaily.com`
    - Development: `http://localhost:4000`
    
    ## Authentication
    All endpoints require JWT authentication unless otherwise specified.
    Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Free tier: 100 requests/minute
    - Premium tier: 500 requests/minute
    - Enterprise tier: 2000 requests/minute
    
    ## Response Format
    All responses follow this structure:
    ```json
    {
      "data": {},
      "error": null,
      "cache": {
        "hit": true,
        "expires_at": "2025-10-20T12:00:00Z"
      }
    }
    ```
    
  contact:
    name: CoinDaily API Support
    email: api@coindaily.com
    url: https://docs.coindaily.com
  license:
    name: Proprietary
    url: https://coindaily.com/license
servers:
  - url: https://api.coindaily.com
    description: Production server
  - url: https://staging-api.coindaily.com
    description: Staging server
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
...
    AITask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agent_id:
          type: string
          format: uuid
        type:
          type: string
          example: "article_generation"
        status:
          type: string
          enum: [queued, processing, completed, failed, cancelled]
        priority:
          type: string
          enum: [low, normal, high, urgent]
        input_data:
          type: object
          additionalProperties: true
        output_data:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        quality_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        cost:
          type: number
          format: float
          example: 0.05
        retry_count:
          type: integer
          example: 0
        max_retries:
          type: integer
          example: 3
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /api/ai/tasks:
    get:
      tags:
        - AI Tasks
      summary: List AI tasks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, processing, completed, failed]
        - name: agent_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AITask'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - AI Tasks
      summary: Create a new AI task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agent_id
                - type
                - input_data
              properties:
                agent_id:
                  type: string
                  format: uuid
                type:
                  type: string
                  example: "article_generation"
                priority:
                  type: string
                  enum: [low, normal, high, urgent]
                  default: normal
                input_data:
                  type: object
                  additionalProperties: true
      responses:
        '201':
          description: Task created successfully
  /api/ai/tasks/{taskId}:
    get:
      tags:
        - AI Tasks
      summary: Get task details
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AITask'
  /api/ai/tasks/{taskId}/cancel:
    post:
      tags:
        - AI Tasks
      summary: Cancel a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task cancelled successfully
  /api/ai/tasks/{taskId}/retry:
    post:
      tags:
        - AI Tasks
      summary: Retry a failed task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task queued for retry

  # ==================== CONTENT PIPELINE ====================
...